online_editor: Fix files from an old project getting added to a new one
When switching from a multi-file demo to another demo while the files
are still getting loaded, the files from the "old" demo were added to
the editor state as they became available.

Fix this by keeping an "edit-era" in the editor that gets incremented
whenever all documents are cleared. Pass this current era into the
callback that handles file loading. Only add files once they are
downloaded if the eras still match up.