Fix panic in GL renderer when PreviewWidget renders too early
If render() is called before the widget is attached, we end up with a
panic in the GL renderer as we cannot extract a gl context out of the
Canvas element.  The element has been created (good), but it's not
hooked into the window's body yet.

Therefore instruction Slint to only render after we've been attached.

In the future this could be generalized to also support detaching.