Clean up online editor startup sequence
Instead of a for loop that awaits ten times one second to see if the lsp
started up (within a microtask that nobody awaits for), resolve the
promise that the editor is ready precisely when we receive the message
from the lsp. In the window's onload handler we can then remove the
loader screen when the editor is ready (as connected microtask).

This means that the other widgets are attached to the DOM in the
meantime and even the preview can being doing its work. To avoid the
loader screen affecting the position of the widgets, place it in a layer
above everything else.