lsp: Deduplicate result of `queryProperties`
When querying properties we walk down the tree of types from the current
elements type down through the base types.

We can see the same element several times that way as it may get
assigned values on the way down to the definition. So de-duplicate
properties in the result and return only the deepest as that is where
the property got defined.