online editor: fix loading of multiple files into the lsp
The preview uses wasm-interpreter's compile_from_string_with_style function,
which takes a closure as parameter that returns the contents of imported files
as string.

The lsp compiled to wasm expects expects a closure in the create() function that returns a Uint8Array,
as that is what vscode.workspace.fs.readFile()
returns in the lsp running inside vs code web.

For the online editor, we returned a string and then tried to interpreter it as Uint8Array,
which doesn't work.

For now, perform the conversion from string to Uint8Array right in the slint/load_file handler.

However, this could also be changed in the future in a way that the closure
for the lsp also expects a string and we change the vs code web lsp build
to decode the Uint8Array into a string.