Ensure that the Qt's NativeStyleMetrics don't crash when called from different threads
The test system may run tests in threads using the fluent style and the
testing backend.  But because the Qt backend is compiled in, the
NativeStyleMetrics is going to be Qt's and it will potentially be
constructed several times in parallel from different thread, causing
warnings and crashes within Qt.
So don't initialize the StyleMetrics, it is only used for the dark style
anyway