online_editor: Use _only_ the uri defined in the ITextModel
We used to have a hand-rolled URL to address the different files open in
the online editor. This got increasingly problematic as the monaco
editor, the slint language server, the previewer and the online editor
code all need to agree on those URLs to be able to cross-reference the
different files.

This patch makes sure we have the URL in ITextModel *only* und use that
consistently everywhere.