online-editor: Improve type safety
Use `monaco.Uri` more in the internal parts of the `EditorWidget`, add
`TextRange` and `TextPosition` types and use them to data around.

Also fix a bug introduced when switching to the tree of items in the
outline editor and another small bug with the file tab name changing for
temporary files. Sorry for mixing this into this patch, but I discover
the issue while working on improving type safety.