femtovg renderer: don't draw svg if they are empty
Otherwise computing the fit_size will lead to NaN and then panic.

The size is empty in the wasm case when loading an html image and it
isn't loaded yet. We will be called again because the htmlimage has an
internal property that will be changed when it is loaded