Properties: fix extra dependency when making a two way binding
This can be reproduced with this example:

```
import { TabWidget } from "std-widgets.slint";
App := Window {
    preferred-width: 700px;
    preferred-height: 500px;
    HorizontalLayout {
        side-bar := TouchArea {
            property <int> current-item: 1;
            clicked => { current-item =  current-item == 0 ? 1 : 0 ; }
        }
        if(side-bar.current-item == 0) :  TabWidget {  }
    }
}
```

With the native style. (click on the window, then resize)
This will result in a stack overflow, because we have a cycle in the
dependency graph and they are marked dirty forever.
The NativeTabWidget will create two way binding internally. And it will
be crated as a result of the layout cache property because that's when
the model will be evaluated and created.  This will inatvertently create
a dependency between the layout cache property, and the two way link.