Correct detect erroneous type-less imports of .slint files
Revert back to the original `ImportTypes` data structure, one per `import ...` "statement" and process individually.

This allows detecting the error and dramatically simplifies the code:
We don't need to do all the dance about unifying
the dependences in a btreemap (just to get
a vector out of it), and we don't need async either.

We still need to collect the dependencies into a vector because build_diagnostics is
borrowed mutably, and that would be done again
later when going through the results.