online editor: Use LspPosition/Range consistently
I was using the editor positions/ranges before, but converting between
LSP and editor values requires having the text at hand as we need to map
from utf-8 to utf-16 based numbering.

This forced all the LSP operations to happen in the editor itself, and
the editor exposing interfaces for the other parts of the online editor
to use. I plan to de-emphasize the editor for the LSP going forward, so I
need to switch out the addressing being used.