vscode: Update version requirements
I just noticed that I had broken the package build before by updating
the types downloaded for vscode. Those types need to be in sync with
the vscode engine version!

Hard-code both to 1.70 for now, I think that is old enough to cover most
installaions of VSCode (which tend to be very new in my experience).