wasm: Force release builds to be space optimized
Unfortunately `wasm-pack` hard-codes the cargo profile used to build, so
I could not just define a new custom profile optimizing for space. I
override the `opt_level` for the `release` profile via the environment
instead.

Unfortunately it is also not easy to set the environment for scripts in
`npm` in a cross-platform way either:-/

So I ended up adding a dependency to `npm` that will read `.env` files to set
up the environment for a process. Set `CARGO_PROFILE_RELEASE_OPT_LEVEL=z` in
that file and only then call `web-pack`.

Before:

-rw-r--r-- 1 dev dev 2066972 Dez 13 18:28 slint_lsp_wasm_bg.wasm
-rw-r--r-- 1 dev dev 5886758 Dez 13 18:32 slint_wasm_interpreter_bg.wasm

After:

-rw-r--r-- 1 dev dev 1376096 Dez 13 18:14 slint_lsp_wasm_bg.wasm
-rw-r--r-- 1 dev dev 4234535 Dez 13 18:08 slint_wasm_interpreter_bg.wasm