Prepare for release: add Cargo.lock and set the version number in the docs
We add the Cargo.lock for the release so that the tagged commit has it,
this will be usefull for the C++ users who build from source with a
tagged version of slint. Because we want to reduce the likelyhod that
they get compilation error in a dependent crate

This will be reverted after the release.