CI: Use opt_level="s" for WASM binaries
This significantly reduces their size (about 40% when I try that
locally), which should reduce download times and traffic from the online
editor.

My first attempt tried to force this for all builds, but failed as we do
not always set up NPM in CI. So this time I focus on the CI only.