lsp: Simplify (left|right)_extend
This is simpler than the previous version and will be easier to
reuse for line-wise deletion, too.

It does sacrifice indention level based (line-) comment splitting, but I
think that will not be missed...

Any comments up to a blank line _before_ a property will be considered to
belong to that property and any comment _after_ the property that starts
without a line break are part of a property now.

There is some extra magic to work when several properties are in the
same line: Basically all comments then refer to the following property.

Update tests accordingly...