Fix crash when using an int model in a repeater with a negative value
Make sure that we return an unsigned for row_count() in C++ and that
we create the usize for the model in Rust capped to zero.

For the interpreter this "worked" by chance as casting a negative floating
number to usize automatically caps at zero, and all values are stored as f64. For safety this patch
applies the same fix though, to be on the safe side.