Fix panic when resetting the model to a smaller model in a ListView
Regresison in 1162ebbb79f819c5fe58ed383cad850bae51e856:
Before that commit, the vew would never have more items than the model
since the model would communicate what items to remove.
But now, it is possible that the model has less elements than the view
currently displays from the previous model, so we need to account for
that.