online editor: Improve waiting for setup to complete
The code used to sometimes throw a strange exception indicating that the
previewer WASM module was not fully up yet when used.

Fix the LspWaiter to really wait for both LSP WASM web worker and WASM
previewer to report back that they are fully up. Previously it only
waited for the Lsp, which surprisingly often works on my machine.