docs: Doc-test files in docs/langref/src
... in addition to the files tested earlier.

This changes the modules generated to include the relative path to the
project root. Otherwise we got errors due to the widgets module getting
defined twice (once in docs/ and once in docs/langref/src).

I think this is saver, as we can end up with files with the same name in
different directories quite easily.