docs: Fix up link to slint language ref in rust docs on slint-ui
This link must work on crates.io and on slint-ui.com. For crates.io we
only have released versions of slint and need a full link to the
language reference of exactly this version (built out of the information
available when running cargo doc).

On slint-ui a relative link works best.

So this patch generates the absolute link needed for crates.io
unconditionally and makes slint-ui update it to a relative link when
publishing the docs.