Fix panic in llr when attempting to access property in base-type in repeater from init code
In the provided test-case, the "title" property in the `Base` type was
optimized away, because binding analysis didn't see that it's used from
the init callback. Consequently it was missing from the property mapping
when lowering to LLR, and the lowering pass for accessing `self.title`
from `init` created a `InNative` property reference instead of `Local`,
because `InNative` is the fallback when the property mapping doesn't
contain it.

This patch makes sure to visit the init code in the property binding
analysis.