online-editor: Add package-lock.json
The current latest versions of all our dependencies do not play well
together.

Monaco-languageclient claims to support monaco-editor 0.34.1 only, so
let's stick to that. We also need to make sure everybody agrees on using
the same vscode-jsonrpc version (which was recently updated in the
vscode dependencies), or the typescript code will fail to communicate
with our language server.

So include the package-lock.json for now, so that we can actually build
something.