online_editor: Do not raise alerts abouts errors opening URLs the user did not explicitly open (#2380)
Do not raise an alert for URLs the LSP/previewer requested. This prevents breaking the flow as the user types an import statement which the LSP/previewer then can not find.

Continue to alert about problems with URLs explicitly opened by the user. This is important so that we do not silently fail when trying to open a file from an URL.

Fixes: #2377