femtovg: speedup TextEdit with many lines
The break_text algorithm was O(N²) in the number of lines because
break_text needs to layout the whole text dspite it would in theory
only need the first line.
Trim the text we pass to break_text so we don't need to re-layout the
end for the text for every call.

This is still O(N²) if there are very long lines.
Not sure we can fix that with the current femtovg break_text api

This makes TextEdit with many lines much faster.  Still a bit slow, but
at least usable