slintpad: Reload if service worker does not come up
Apparently browsers do not set a controlling service worker when
force-reloading a page. Work around this by reloading the page.

We run into this issue only while the load screen is still up. So we
have no state to take care of or anything, so go for the simplest
thinkable "fix": Reload the page.