LSP: don't throw a protocol error when a symbol can't be renamed
I misread the spec that we should return an error if there is nothing
to rename. We should just return None instead.

This means that the editor will no longer show an error message when
this happens (and also that we can rename stuff in rust code without
getting annoying error)

Technicly, we could still return an error if the user tries to rename a
property or so, but i don't think its worth it