LSP: fix error location
Commit 893983e8d320bd248c47f3d57292d432ec4ded69 fixed Diagnostic::line_column to
make the column 0-based instead of 1-based. But the LSP was not adjusted yet.
This makes sure the error is reported in the right location

Also adjust a couple of more place.
The C++ slint::interpreter::Diagnostic documentation was changed to be consistent with
the Rust documentation and new behavior