LSP: Add code lens to toggle Design Mode
Add a toggle command to the LSP for this effect. This is necessary as
the LSP provides the code lens information and has no idea of the state
stored in the preview.

Keep the existing set_design_mode commands for Slintpad, as it is easier
for the state stored on the web site and in the WASM side to go out of
sync. This is not an issue in VS Code at this point.