LSP: Add `toggleDesignMode` command
Add a toggle command to the LS and use them in VSCode.

Keep the existing set_design_mode commands for Slintpad, as it is easier
for the state stored on the web site and in the WASM side to go out of
sync. This is not an issue in VS Code at this point.