Disable winit accessor tests
Now that we always create a window, we end up also creating an event
loop, and that's inherently incompatible with the rust test harness that
creates threads.

We can try to re-enable this later on platforms where we can run an
event loop in a thread. I have a local WIP patch for that.