textlayout: Fix missing byte range on empty lines
Even for an empty line, we should report a byte range so that clicking into an empty line
can produce a cursor offset.

The previous behavior was covered by an existing test, that's now adjusted.
In addition a new test covers that we report the correct
start also for lines starting with whitespace (non-breaking, that is - regular
breaking whitespace is covered by the line breaker).