Turn `RepeatedElementInfo` into an enum
... that can distinguish between a Repeater (`if`/`for`) and an Embedding
(to be added `Embed` item).

This patch _should_ not change behavior for either "no repeater" or "repeater"
cases. The "embedding" case is currently a `todo!()` as embedding is not
implemented yet.

In some places we definitely and obviously do not care for embedding (e.g.
when counting repeaters) and then the code is simplified with that in mind.